Rename files, move LaunchMenu details to it's own file / struct.
